--- csharp-level1-les01.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <RootNamespace>csharp_level1_les01</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Include=".github\PULL_REQUEST_TEMPLATE.md" />
-  </ItemGroup>
-
-</Project>
