--- csharp-level1-les01/lession01.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-    <RootNamespace>csharp_level1</RootNamespace>
-  </PropertyGroup>
-
-</Project>
