+++ csharp-level1.sln
@@ -3,17 +3,17 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29519.87
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "lession01", "csharp-level1-les01\lession01.csproj", "{E53E4641-9F1F-41D5-B291-043B99E87EA0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "lession02", "lession02\lession02.csproj", "{F0631EA2-EF80-4705-B903-F161D3E2DBDC}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lession03", "lession03", "{CA05B2E1-D851-45B8-A8EB-DB9E32E54D8A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "lession03-t1", "lession03-t1\lession03-t1.csproj", "{55A8374A-EAB5-41F1-9DBF-29F8A346F10E}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lession01", "lession01", "{35210588-BD6D-4585-BC0F-D58738B7D8C4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "lession02", "lession02\lession02.csproj", "{F0631EA2-EF80-4705-B903-F161D3E2DBDC}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lession02", "lession02", "{E9F45C4E-B0B4-4699-B6A8-C3C180FD0970}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "lession01", "lession01\lession01.csproj", "{A93D7F94-6E6B-4C3E-8583-91FEF59B9D68}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lession01", "lession01", "{458E59EE-2DF7-4CF9-9FAE-1118DEBDEF07}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "lession03", "lession03", "{C317019B-AD48-4013-946C-74C1ADB6B06F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "lession03", "lession03\lession03.csproj", "{1994ED12-08FB-417B-B845-F838BD90D5EE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,26 +21,26 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E53E4641-9F1F-41D5-B291-043B99E87EA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E53E4641-9F1F-41D5-B291-043B99E87EA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E53E4641-9F1F-41D5-B291-043B99E87EA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E53E4641-9F1F-41D5-B291-043B99E87EA0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F0631EA2-EF80-4705-B903-F161D3E2DBDC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F0631EA2-EF80-4705-B903-F161D3E2DBDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F0631EA2-EF80-4705-B903-F161D3E2DBDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F0631EA2-EF80-4705-B903-F161D3E2DBDC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{55A8374A-EAB5-41F1-9DBF-29F8A346F10E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{55A8374A-EAB5-41F1-9DBF-29F8A346F10E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{55A8374A-EAB5-41F1-9DBF-29F8A346F10E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{55A8374A-EAB5-41F1-9DBF-29F8A346F10E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A93D7F94-6E6B-4C3E-8583-91FEF59B9D68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A93D7F94-6E6B-4C3E-8583-91FEF59B9D68}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A93D7F94-6E6B-4C3E-8583-91FEF59B9D68}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A93D7F94-6E6B-4C3E-8583-91FEF59B9D68}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1994ED12-08FB-417B-B845-F838BD90D5EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1994ED12-08FB-417B-B845-F838BD90D5EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1994ED12-08FB-417B-B845-F838BD90D5EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1994ED12-08FB-417B-B845-F838BD90D5EE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{E53E4641-9F1F-41D5-B291-043B99E87EA0} = {35210588-BD6D-4585-BC0F-D58738B7D8C4}
 		{F0631EA2-EF80-4705-B903-F161D3E2DBDC} = {E9F45C4E-B0B4-4699-B6A8-C3C180FD0970}
-		{55A8374A-EAB5-41F1-9DBF-29F8A346F10E} = {CA05B2E1-D851-45B8-A8EB-DB9E32E54D8A}
+		{A93D7F94-6E6B-4C3E-8583-91FEF59B9D68} = {458E59EE-2DF7-4CF9-9FAE-1118DEBDEF07}
+		{1994ED12-08FB-417B-B845-F838BD90D5EE} = {C317019B-AD48-4013-946C-74C1ADB6B06F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4A56D6BC-D6E9-4D3A-B237-56B434F9C964}
