+++ lession03/lession03.csproj
@@ -3,7 +3,6 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.0</TargetFramework>
-    <RootNamespace>lession03_t1</RootNamespace>
   </PropertyGroup>
 
 </Project>
